move component editors in separate package
