+++ kool-editor/src/commonMain/kotlin/de/fabmax/kool/editor/ui/componenteditors/JointEditor.kt
@@ -1,4 +1,4 @@
-package de.fabmax.kool.editor.ui
+package de.fabmax.kool.editor.ui.componenteditors
 
 import de.fabmax.kool.editor.actions.EditorAction
 import de.fabmax.kool.editor.actions.SetComponentDataAction
@@ -7,6 +7,7 @@ import de.fabmax.kool.editor.components.JointComponent
 import de.fabmax.kool.editor.components.RigidActorComponent
 import de.fabmax.kool.editor.data.EntityId
 import de.fabmax.kool.editor.data.JointData
+import de.fabmax.kool.editor.ui.*
 import de.fabmax.kool.modules.ui2.*
 
 class JointEditor : ComponentEditor<JointComponent>() {
