+++ kool-editor/src/commonMain/kotlin/de/fabmax/kool/editor/ui/componenteditors/MaterialAttributeSourcePopup.kt
@@ -1,8 +1,9 @@
-package de.fabmax.kool.editor.ui
+package de.fabmax.kool.editor.ui.componenteditors
 
 import de.fabmax.kool.editor.AssetItem
 import de.fabmax.kool.editor.KoolEditor
 import de.fabmax.kool.editor.data.*
+import de.fabmax.kool.editor.ui.*
 import de.fabmax.kool.modules.ui2.*
 import de.fabmax.kool.pipeline.Attribute
 import de.fabmax.kool.util.Color
