+++ kool-editor/src/commonMain/kotlin/de/fabmax/kool/editor/ui/componenteditors/PbrSplatMaterialEditor.kt
@@ -1,10 +1,11 @@
-package de.fabmax.kool.editor.ui
+package de.fabmax.kool.editor.ui.componenteditors
 
 import de.fabmax.kool.editor.KoolEditor
 import de.fabmax.kool.editor.actions.SetComponentDataAction
 import de.fabmax.kool.editor.api.AssetReference
 import de.fabmax.kool.editor.components.MaterialComponent
 import de.fabmax.kool.editor.data.*
+import de.fabmax.kool.editor.ui.*
 import de.fabmax.kool.modules.ksl.KslPbrSplatShader
 import de.fabmax.kool.modules.ui2.*
 import de.fabmax.kool.util.Color
