+++ kool-editor/src/commonMain/kotlin/de/fabmax/kool/editor/ui/componenteditors/ScenePropertiesEditor.kt
@@ -1,9 +1,12 @@
-package de.fabmax.kool.editor.ui
+package de.fabmax.kool.editor.ui.componenteditors
 
 import de.fabmax.kool.editor.actions.SetComponentDataAction
 import de.fabmax.kool.editor.components.CameraComponent
 import de.fabmax.kool.editor.components.SceneComponent
 import de.fabmax.kool.editor.data.EntityId
+import de.fabmax.kool.editor.ui.Icons
+import de.fabmax.kool.editor.ui.labeledCombobox
+import de.fabmax.kool.editor.ui.labeledIntTextField
 import de.fabmax.kool.modules.ksl.blocks.ToneMapping
 import de.fabmax.kool.modules.ui2.UiScope
