+++ kool-editor/src/commonMain/kotlin/de/fabmax/kool/editor/ui/componenteditors/TransformEditor.kt
@@ -1,4 +1,4 @@
-package de.fabmax.kool.editor.ui
+package de.fabmax.kool.editor.ui.componenteditors
 
 import de.fabmax.kool.editor.KoolEditor
 import de.fabmax.kool.editor.actions.FusedAction
@@ -10,6 +10,7 @@ import de.fabmax.kool.editor.components.localToGlobalD
 import de.fabmax.kool.editor.data.TransformData
 import de.fabmax.kool.editor.data.Vec3Data
 import de.fabmax.kool.editor.data.Vec4Data
+import de.fabmax.kool.editor.ui.*
 import de.fabmax.kool.math.*
 import de.fabmax.kool.modules.gizmo.GizmoFrame
 import de.fabmax.kool.modules.ui2.*
